requirements.txt
yfinance==0.2.63
newsapi-python==0.2.7
pandas==2.3.0
numpy==2.2.2
ta==0.11.0
backtrader==1.9.78.123
requests==2.31.0
tradingview-ta==3.3.0
tinkoff-investments==0.2.0b100
asyncio==3.4.3
python-dotenv==1.1.0
